drizzlepac/hlautils/svm_quality_analysis.py: - moved subroutine hapsequencer.run_sourcelist_comparision() to svm_quality_analysis.py. - renamed subroutine svm_quality_analysis.run_sourcelist_comparision() to svm_quality_analysis.run_hla_sourcelist_comparision() - moved subroutine hapsequencer.correct_hla_classic_ra_dec() to svm_quality_analysis.py. - moved associated import statements over as well - updated impacted subroutine calls. - updated svm_quality_analysis.py to facilitate execution of run_hla_sourcelist_comparision() from either the command line or as a python subroutine call.
